requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic<2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.3.1
python-multipart>=0.0.6
tenacity>=8.2.2
langchain>=0.0.200
google-generativeai>=0.3.0
